add grpcio 1.57.0 pin to provided helm chart images
Summary:
Attempt to resolve reported memory leak versions with . Apparently the underlying issue is also resolved in python 3.13, so once we officially support that in dagster, we could potentially move these images to 3.13 and remove this.

Test Plan:
Deploy the helm chart locally with these built images

Changelog:
Added a pin to the grpcio version  used by the Dagster Helm chart daemon and webserver images, to a version that has been reported to avoid memory leaks.